refactor(package-manifest): extract inline test module into dedicated file
Move the inline `#[cfg(test)] mod tests { ... }` block (190 lines)
from `lib.rs` into `tests.rs` following CODE_STYLE_GUIDE.md. No
behavioral change.

Part of #358.
